Move login/register form to same map
